logOut done

+++ src/app/components/navbar/navbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { AuthService } from '../../service/auth.service';
 
 @Component({
   selector: 'app-navbar',
@@ -9,7 +10,7 @@ import { ToastrService } from 'ngx-toastr';
 })
 export class NavbarComponent implements OnInit {
 
-  constructor(private router: Router, private toastr: ToastrService) { }
+  constructor(private router: Router, private toastr: ToastrService, private authService: AuthService) { }
 
   ngOnInit() {
   }
@@ -19,10 +20,14 @@ export class NavbarComponent implements OnInit {
   }
 
   onLogOut() {
-    const username = localStorage.getItem('username');
-    localStorage.clear();
-    this.router.navigate(['/signin']);
-    this.showInfoToast('User ' + username + ' has logged out');
+    if (!this.authService.isTokenExpired()) {
+      const username = localStorage.getItem('username');
+      localStorage.clear();
+      this.router.navigate(['/signin']);
+      this.showInfoToast('User ' + username + ' has logged out');
+    }
+    
+    else this.showInfoToast('Please log in')
   }
 
 }
